Melhoras no audit e em ocirrencias e melhorias no envios das senhas dos email do morador

+++ database/migrations/2024_10_27_161406_add_column_status_resident.php
@@ -13,9 +13,10 @@ return new class extends Migration
     {
         //
         Schema::table('resident', function (Blueprint $table) {
-            $table->integer('status_id')->nullable();
+            $table->unsignedBigInteger('status_id')->nullable();
+            // $table->integer('status_id')->nullable();
             // $table->foreignId('status_id')->constrained('status');
-           
+
         });
     }
